+++ lib/service/translate/microsoft.dart
@@ -4,14 +4,19 @@ import 'package:anx_reader/enums/lang_list.dart';
 import 'package:anx_reader/service/translate/index.dart';
 import 'package:anx_reader/utils/log/common.dart';
 import 'package:dio/dio.dart';
+import 'package:flutter/material.dart';
 
 const urlMicrosoft =
     'https://api-edge.cognitive.microsofttranslator.com/translate';
 const urlMicrosoftAuth = 'https://edge.microsoft.com/translate/auth';
 
-class MicrosoftTranslateProvider implements TranslateServiceProvider {
+class MicrosoftTranslateProvider extends TranslateServiceProvider {
   @override
-  Stream<String> translate(
+  Widget translate(String text, LangListEnum from, LangListEnum to) {
+    return convertStreamToWidget(translateStream(text, from, to));
+  }
+
+  Stream<String> translateStream(
       String text, LangListEnum from, LangListEnum to) async* {
     try {
       yield "...";